feat: migrate server.js entry point to TypeScript
Complete Phase 6 of TypeScript migration:
- Create src/server.ts with full TypeScript implementation
- Convert server.js to a thin wrapper that loads compiled TS
- Export ConvertImageOptions and Spectra6ColorWithLab types
- Fix MCP module path for compiled output

All 220 tests passing. TypeScript migration complete!
